+++ src/util.ts
@@ -116,6 +116,8 @@ export function getProjectAddress(project: string): {
   }
 }
 
-export function getVersion(): string {
-  return argon.version().replace("argon-rbx ", "").trim()
+export function getVersion(): string | undefined {
+  try {
+    return argon.version().replace("argon-rbx ", "").trim()
+  } catch {}
 }